File store no longer escapes
Leaps used to mandate that documents had a full title and description,
this was never necessary and rightly removed. However, the old file
store behaviour was to store both the title and description along with
the content of a document to a file, so it had to be a line delimited
format of escaped values.

Obviously now this can be changed to be a standard file editing
mechanic, where you can open and create regular text files and everyone
wins.