New file_authenticator, trimmed some fat.
The new file_authenticator is a type of authentication designed to work
with the file document store option. Its core job is to ensure that any
files being accessed through leaps are found within a particular
directory.

The authenticator also allows you to register an HTTP endpoint for
GETing a list of available files for editing, this will allow you to
write a client that lists those files and connects to them without
introducing another service for file walking.