Write a client for folder sharing
Leaps now has the capability to run as a local folder sharing service.
You can run it in a directory and the files in that folder can instantly
be collaboratively edited by anyone able to connect to your bound
address.

The share_dir client within the static directory is an interface that
uses the ACE editor to work with the service in this configuration. Once
finished I'll write a how-to guide for setting up the config etc.