Add session ID to binder clients
It's important that a user connected to a leaps server is able to
distinguish messages coming from each connection and not just each user,
since these messages can be used to send state such as cursor positions
which are unique per user per connection.

This change adds a new sessionID to each client connection to a binder,
so that now when a client gets a message they are able to read both a
user ID and a session ID to identify the specific session.

This commit changes the API and therefore the client and documentation
need to be updated.